requirements.txt
pdf2image==1.16.2
Pillow==9.4.0
pytesseract==0.3.10
streamlit==1.18.1
